+++ pkg/sonar/scan.go
@@ -1,7 +1,10 @@
 package sonar
 
 import (
+	"bufio"
 	"fmt"
+	"os"
+	"strings"
 
 	"github.com/opendevstack/pipeline/internal/command"
 )
@@ -11,6 +14,23 @@ type PullRequest struct {
 	Branch string
 	Base   string
 }
+
+// Scan report
+type ReportTask struct {
+	ProjectKey    string
+	ServerUrl     string
+	ServerVersion string
+	Branch        string
+	DashboardUrl  string
+	CeTaskId      string
+	CeTaskUrl     string
+}
+
+const (
+	ScannerworkDir     = ".scannerwork"
+	ReportTaskFilename = "report-task.txt"
+	ReportTaskFile     = ScannerworkDir + "/" + ReportTaskFilename
+)
 
 // Scan scans the source code and uploads the analysis to given SonarQube project.
 // If pr is non-nil, information for pull request decoration is sent.
@@ -25,6 +45,10 @@ func (c *Client) Scan(sonarProject, branch, commit string, pr *PullRequest) (str
 	if c.clientConfig.Debug {
 		scannerParams = append(scannerParams, "-X")
 	}
+	// Both Branch Analysis and Pull Request Analysis are only available
+	// starting in Developer Edition, see
+	// https://docs.sonarqube.org/latest/branches/overview/ and
+	// https://docs.sonarqube.org/latest/analysis/pull-request/.
 	if c.clientConfig.ServerEdition != "community" {
 		if pr != nil {
 			scannerParams = append(
@@ -45,4 +69,37 @@ func (c *Client) Scan(sonarProject, branch, commit string, pr *PullRequest) (str
 		return string(stdout), fmt.Errorf("scanning failed: %w, stderr: %s", err, string(stderr))
 	}
 	return string(stdout), nil
+}
+
+/*
+Example of the file located in .scannerwork/report-task.txt:
+	projectKey=XXXX-python
+	serverUrl=https://sonarqube-ods.XXXX.com
+	serverVersion=8.2.0.32929
+	branch=dummy
+	dashboardUrl=https://sonarqube-ods.XXXX.com/dashboard?id=XXXX-python&branch=dummy
+	ceTaskId=AXxaAoUSsjAMlIY9kNmn
+	ceTaskUrl=https://sonarqube-ods.XXXX.com/api/ce/task?id=AXxaAoUSsjAMlIY9kNmn
+*/
+func (c *Client) ExtractComputeEngineTaskID(filename string) (string, error) {
+	file, err := os.Open(filename)
+	if err != nil {
+		return "", err
+	}
+	defer file.Close()
+
+	taskIDPrefix := "ceTaskId="
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		line := strings.TrimSpace(scanner.Text())
+		if strings.HasPrefix(line, taskIDPrefix) {
+			return strings.TrimPrefix(line, taskIDPrefix), nil
+		}
+	}
+
+	if err := scanner.Err(); err != nil {
+		return "", err
+	}
+
+	return "", fmt.Errorf("properties file %s does not contain %s", filename, taskIDPrefix)
 }